fix(page-settings-and-site-router/site-service): 1. Add logic for disabling button 2. Add logic to fetch previous notification upon initial settings page load 3. Modified site router to have ability to remove existing notification through the setNotification endpoint with empty string
